feat: Start adding ability to force v1 polyfill.
Separated out the non-native to v1 polyfill into a function and used that in index.js.

resolves #106
